CDE告警处理 20260209

+++ frameworks/kits/js/common/image_source_napi.cpp
@@ -3718,8 +3718,8 @@ static std::unique_ptr<ImageSourceAsyncContext> UnwrapContextForReadImageMetadat
         uint32_t errorCode;
         napi_get_value_int32(env, argValue[NUM_1], &index);
         uint32_t frameCount = context->rImageSource->GetFrameCount(errorCode);
-        IMG_NAPI_CHECK_RET_D((errorCode == SUCCESS) && (index >= 0) && (index < frameCount), nullptr,
-            IMAGE_LOGE("Invalid readImageMetadata index"));
+        IMG_NAPI_CHECK_RET_D((errorCode == SUCCESS) && (index >= 0) &&
+            (static_cast<uint32_t>(index) < frameCount), nullptr, IMAGE_LOGE("Invalid readImageMetadata index"));
         context->index = static_cast<uint32_t>(index);
         IMAGE_LOGD("index is %{public}d", index);
     }
